requirements.txt
flask
flask-cors
pyngrok
google-generativeai
transformers
huggingface-hub
python-dotenv
PyTorch